Move daemon to bin/glusterd2/daemon
