+++ bin/glusterd2/commands/volumes/brickutils.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/gluster/glusterd2/bin/glusterd2/brick"
-	"github.com/gluster/glusterd2/daemon"
+	"github.com/gluster/glusterd2/bin/glusterd2/daemon"
 	"github.com/gluster/glusterd2/pkg/utils"
 
 	"github.com/pborman/uuid"
