+++ bin/glusterd2/commands/volumes/volume-status.go
@@ -5,12 +5,12 @@ import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/bin/glusterd2/brick"
+	"github.com/gluster/glusterd2/bin/glusterd2/daemon"
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/bin/glusterd2/pmap"
 	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
 	"github.com/gluster/glusterd2/bin/glusterd2/volume"
-	"github.com/gluster/glusterd2/daemon"
 	"github.com/gluster/glusterd2/pkg/errors"
 
 	"github.com/gorilla/mux"
