+++ bin/glusterd2/commands/volumes/volume-stop.go
@@ -4,11 +4,11 @@ import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/bin/glusterd2/brick"
+	"github.com/gluster/glusterd2/bin/glusterd2/daemon"
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	restutils "github.com/gluster/glusterd2/bin/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
 	"github.com/gluster/glusterd2/bin/glusterd2/volume"
-	"github.com/gluster/glusterd2/daemon"
 	"github.com/gluster/glusterd2/pkg/errors"
 
 	"github.com/gorilla/mux"
