+++ bin/glusterd2/daemon/connection.go
