+++ bin/glusterd2/daemon/daemon.go
