+++ bin/glusterd2/daemon/store.go
