+++ bin/glusterd2/daemon/storeddaemon.go
