+++ bin/glusterd2/main.go
@@ -7,13 +7,13 @@ import (
 	"strings"
 	"time"
 
+	"github.com/gluster/glusterd2/bin/glusterd2/daemon"
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/bin/glusterd2/peer"
 	"github.com/gluster/glusterd2/bin/glusterd2/servers"
 	"github.com/gluster/glusterd2/bin/glusterd2/store"
 	"github.com/gluster/glusterd2/bin/glusterd2/volgen"
 	"github.com/gluster/glusterd2/bin/glusterd2/xlator"
-	"github.com/gluster/glusterd2/daemon"
 	"github.com/gluster/glusterd2/pkg/utils"
 	"github.com/gluster/glusterd2/version"
